메인페이지 상품 리뷰건수, 평점 출력

+++ src/main/java/com/example/MCM/domain/review/ReviewCreateDTO.java
@@ -16,5 +16,5 @@ public class ReviewCreateDTO {
     private String content;
 
     @NotNull(message = "별점을 선택해 주세요.")
-    private Integer starScore;
+    private double starScore;
 }
